# docs/architecture/openapi.snapshot.json
{
  "components": {
    "schemas": {
...
      "DailyReportResponse": {
        "properties": {
          "avg_focus_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avg Focus Score"
          },
          "avg_posture_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avg Posture Score"
          },
          "best_hours": {
            "items": {
              "type": "string"
            },
            "title": "Best Hours",
            "type": "array"
          },
          "break_count": {
            "title": "Break Count",
            "type": "integer"
          },
          "date": {
            "format": "date",
            "title": "Date",
            "type": "string"
          },
          "distraction_count": {
            "title": "Distraction Count",
            "type": "integer"
          },
          "longest_break_seconds": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Longest Break Seconds"
          },
          "timeline": {
            "items": {
              "$ref": "#/components/schemas/TimelineEntryResponse"
            },
            "title": "Timeline",
            "type": "array"
          },
          "worst_hours": {
            "items": {
              "type": "string"
            },
            "title": "Worst Hours",
            "type": "array"
          }
        },
        "required": [
          "date",
          "distraction_count",
          "break_count",
          "best_hours",
          "worst_hours",
          "timeline"
        ],
        "title": "DailyReportResponse",
        "type": "object"
      },
...
      "EventType": {
        "enum": [
          "session_started",
          "session_ended",
          "session_paused",
          "session_resumed",
          "looking_at_screen",
          "looked_away",
          "phone_detected",
          "phone_pickup",
          "phone_down",
          "good_posture",
          "poor_posture",
          "slump_started",
          "slump_ended",
          "yawn",
          "long_blink",
          "face_touch",
          "fatigue_level_changed",
          "left_desk",
          "returned",
          "coffee_drink",
          "focus_score_updated",
          "camera_disconnected",
          "camera_reconnected",
          "low_confidence_suppressed"
        ],
        "title": "EventType",
        "type": "string"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ReportStatusResponse": {
        "properties": {
          "download_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Download Url"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status"
        ],
        "title": "ReportStatusResponse",
        "type": "object"
      },
...
      "TimelineEntryResponse": {
        "properties": {
          "confidence": {
            "title": "Confidence",
            "type": "number"
          },
          "event_type": {
            "$ref": "#/components/schemas/EventType"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "time": {
            "title": "Time",
            "type": "string"
          }
        },
        "required": [
          "time",
          "label",
          "event_type",
          "confidence"
        ],
        "title": "TimelineEntryResponse",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "WeeklyReportResponse": {
        "properties": {
          "avg_focus_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avg Focus Score"
          },
          "avg_posture_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avg Posture Score"
          },
          "best_hours": {
            "items": {
              "type": "string"
            },
            "title": "Best Hours",
            "type": "array"
          },
          "break_count": {
            "title": "Break Count",
            "type": "integer"
          },
          "daily_breakdown": {
            "items": {
              "$ref": "#/components/schemas/DailyReportResponse"
            },
            "title": "Daily Breakdown",
            "type": "array"
          },
          "distraction_count": {
            "title": "Distraction Count",
            "type": "integer"
          },
          "longest_break_seconds": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Longest Break Seconds"
          },
          "week_end": {
            "format": "date",
            "title": "Week End",
            "type": "string"
          },
          "week_start": {
            "format": "date",
            "title": "Week Start",
            "type": "string"
          },
          "worst_hours": {
            "items": {
              "type": "string"
            },
            "title": "Worst Hours",
            "type": "array"
          }
        },
        "required": [
          "week_start",
          "week_end",
          "distraction_count",
          "break_count",
          "best_hours",
          "worst_hours",
          "daily_breakdown"
        ],
        "title": "WeeklyReportResponse",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Local HTTP/WebSocket API for the Attune desktop app \u2014 session lifecycle, events, analytics reports, settings, and export. See docs/architecture/06-api-specification.md.",
    "title": "Attune API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/reports/daily": {
      "get": {
        "operationId": "get_daily_report_api_v1_reports_daily_get",
        "parameters": [
          {
            "in": "query",
            "name": "date",
            "required": true,
            "schema": {
              "format": "date",
              "title": "Date",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DailyReportResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Daily Report",
        "tags": [
          "reports"
        ]
      }
    },
    "/api/v1/reports/weekly": {
      "get": {
        "operationId": "get_weekly_report_api_v1_reports_weekly_get",
        "parameters": [
          {
            "in": "query",
            "name": "week_start",
            "required": true,
            "schema": {
              "format": "date",
              "title": "Week Start",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WeeklyReportResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Weekly Report",
        "tags": [
          "reports"
        ]
      }
    },
    "/api/v1/reports/{report_id}": {
      "get": {
        "description": "Polling endpoint for the async PDF/PNG jobs POST /export creates.",
        "operationId": "get_report_status_api_v1_reports__report_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "report_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Report Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportStatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Report Status",
        "tags": [
          "reports"
        ]
      }
    },
    "/api/v1/reports/{report_id}/download": {
      "get": {
        "operationId": "download_report_api_v1_reports__report_id__download_get",
        "parameters": [
          {
            "in": "path",
            "name": "report_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Report Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Download Report",
        "tags": [
          "reports"
        ]
      }
    },
...
  }
}